feat: Voting for Comments
I added a little something to VoteMechanism component to update
using the two different API calls, based on the component type param
passed in. Also, in the meantime, I had to refactor how the comments
are saved in the store. Now, on every new Post load, its comments are
loaded into the store, as usual. However they do not replace the old
comments. They are added to it. The CommentList component then only
shows those with the right parentId. Nice!